Fixed victim servo bug. (Time should be unsigned long, not byte!)

+++ Code/Libraries/Motors/Motors.h
@@ -21,8 +21,8 @@ struct MotorConfig
 	byte victim_servo_open_angle;
 	//byte right_servo_open_angle;
 
-	byte servo_close_time;
-	byte servo_open_time;
+	unsigned long servo_close_time;
+	unsigned long servo_open_time;
 };
 
 /**
